# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  POS-fiplex_internal_common.ErrorResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
    type: object
  POS-fiplex_internal_common.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
  POS-fiplex_internal_common_pagination.Pagination:
    properties:
      current_page:
        type: integer
      per_page:
        type: integer
      total_data:
        type: integer
      total_page:
        type: integer
    type: object
...
  POS-fiplex_internal_user_repository.UserRole:
    enum:
    - admin
    - cashier
    - manager
    type: string
    x-enum-varnames:
    - UserRoleAdmin
    - UserRoleCashier
    - UserRoleManager
...
  fiber.Map:
    additionalProperties: {}
    type: object
...
  internal_user.CreateUserRequest:
    properties:
      email:
        maxLength: 100
        type: string
      is_active:
        type: boolean
      password:
        maxLength: 100
        minLength: 6
        type: string
      role:
        allOf:
        - $ref: '#/definitions/POS-fiplex_internal_user_repository.UserRole'
        enum:
        - admin
        - manager
        - cashier
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - role
    - username
    type: object
  internal_user.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  internal_user.LoginResponse:
    properties:
      Token:
        type: string
      expired_at:
        type: string
      profile:
        $ref: '#/definitions/internal_user.ProfileResponse'
      refresh_token:
        type: string
    type: object
  internal_user.ProfileResponse:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      role:
        $ref: '#/definitions/POS-fiplex_internal_user_repository.UserRole'
      updated_at:
        type: string
      username:
        type: string
    type: object
  internal_user.RegisterRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 32
        minLength: 8
        type: string
      role:
        $ref: '#/definitions/POS-fiplex_internal_user_repository.UserRole'
      username:
        maxLength: 32
        minLength: 3
        type: string
    required:
    - email
    - password
    - role
    - username
    type: object
  internal_user.UpdatePasswordRequest:
    properties:
      new_password:
        maxLength: 32
        minLength: 8
        type: string
      old_password:
        maxLength: 32
        minLength: 8
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  internal_user.UsersResponse:
    properties:
      pagination:
        $ref: '#/definitions/POS-fiplex_internal_common_pagination.Pagination'
      users:
        items:
          $ref: '#/definitions/internal_user.ProfileResponse'
        type: array
    type: object
host: localhost:8080
info:
  contact: {}
  description: POS Fiplex API
  title: POS Fiplex API
  version: "1.0"
paths:
...
  /auth/add:
    post:
      consumes:
      - application/json
      description: 'Register a new user with a specific role (Roles: admin)'
      parameters:
      - description: New user details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_user.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User added successfully
          schema:
            allOf:
            - $ref: '#/definitions/POS-fiplex_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_user.ProfileResponse'
              type: object
        "400":
          description: Invalid request body or validation failed
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "403":
          description: Forbidden - higher role assignment attempt
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "409":
          description: User, username, or email already exists
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
      summary: Add new user
      tags:
      - Auth
      x-roles:
      - admin
  /auth/login:
    post:
      consumes:
      - application/json
      description: 'Authenticate user and return access/refresh tokens via cookies
        and response body (Roles: public)'
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Authenticated successfully
          schema:
            allOf:
            - $ref: '#/definitions/POS-fiplex_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_user.LoginResponse'
              type: object
        "400":
          description: Invalid request body or validation failed
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "401":
          description: Invalid username or password
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
      summary: Login
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: 'Clear access and refresh token cookies (Roles: authenticated)'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged out
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
      summary: Logout
      tags:
      - Auth
      x-roles:
      - admin
      - manager
      - cashier
  /auth/me:
    get:
      consumes:
      - application/json
      description: 'Get detailed profile information for the authenticated user session
        (Roles: authenticated)'
      produces:
      - application/json
      responses:
        "200":
          description: Profile retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/POS-fiplex_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_user.ProfileResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
      summary: Get current profile
      tags:
      - Auth
      x-roles:
      - admin
      - manager
      - cashier
  /auth/me/avatar:
    put:
      consumes:
      - multipart/form-data
      description: 'Upload and update the profile picture for the current user (Roles:
        authenticated)'
      parameters:
      - description: Avatar image file
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Avatar updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/POS-fiplex_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_user.ProfileResponse'
              type: object
        "400":
          description: Invalid file or dimensions
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
      summary: Update avatar
      tags:
      - Auth
      x-roles:
      - admin
      - manager
      - cashier
  /auth/me/password:
    put:
      consumes:
      - application/json
      description: 'Update the password for the current user session (Roles: authenticated)'
      parameters:
      - description: Password update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_user.UpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password updated successfully
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.SuccessResponse'
        "400":
          description: Invalid request body or validation failed
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
      summary: Update password
      tags:
      - Auth
      x-roles:
      - admin
      - manager
      - cashier
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: 'Issue a new access token using a valid refresh token cookie (Roles:
        public/authenticated)'
      produces:
      - application/json
      responses:
        "200":
          description: Token refreshed successfully
          schema:
            allOf:
            - $ref: '#/definitions/POS-fiplex_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/fiber.Map'
              type: object
        "401":
          description: Invalid or expired session
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
      summary: Refresh token
      tags:
      - Auth
...
  /users:
    get:
      consumes:
      - application/json
      description: 'Retrieve a list of users with pagination, filtering, and sorting
        (Roles: admin, manager, cashier)'
      parameters:
      - default: 1
        description: Page number (default 1)
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page (default 10)
        in: query
        name: limit
        type: integer
      - description: Search by username or email
        in: query
        name: search
        type: string
      - description: Filter by User Role
        enum:
        - admin
        - cashier
        - manager
        in: query
        name: role
        type: string
      - description: Filter by Active Status
        in: query
        name: is_active
        type: boolean
      - description: Filter by Account Status
        enum:
        - active
        - deleted
        - all
        in: query
        name: status
        type: string
      - description: Sort by column
        enum:
        - created_at
        - username
        in: query
        name: sortBy
        type: string
      - description: Sort direction
        enum:
        - asc
        - desc
        in: query
        name: sortOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Users retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/POS-fiplex_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_user.UsersResponse'
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
      summary: Get all users
      tags:
      - Users
      x-roles:
      - admin
      - manager
      - cashier
    post:
      consumes:
      - application/json
      description: 'Create a new user account (Roles: admin)'
      parameters:
      - description: New user details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/internal_user.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            allOf:
            - $ref: '#/definitions/POS-fiplex_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_user.ProfileResponse'
              type: object
        "400":
          description: Invalid request body or validation failed
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "409":
          description: User, username, or email already exists
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
      summary: Create user
      tags:
      - Users
      x-roles:
      - admin
...
swagger: "2.0"